Handle non-ClassType's in `Clone`.
The IR spec allows any subtype of `jl.Cloneable` as argument to
the `Clone` node. In particular, that includes array types in
addition to `ClassType`s.

Our library+codegen is written in a way such that these cases do
not happen in practice, but the IR spec still allows it. Once we
enable the optimizer, we will actually get those cases in practice
as well.