Introduce Memory Management API for WebAssembly backend
This commit introduces a memory management API for WebAssembly, enabling direct manipulation of memory segments on Wasm linear memory.

`MemoryAllocator` handles the allocation and freeing of memory, while `MemorySegment` allows for read/write data on the allocated memory.
Those APIs are implemented as a intrinsic functions: Wasm linker backend
translates the method body of `allocate`, `free` of `MemoryAllocator`
and `_storeX`, `_loadX` of `MemorySegment` to `Transient.Value` and then
transformed into Wasm via `FunctionEmitter`.

We use `Transient.Value` instead of defining new `Tree`s in SJSIR,
so that we wouldn't introduce the breaking change once we need to modify
the structure of those structures.
We'll keep them `Transient.Value` at least until we make sure we found
the "right" IR structure.