Opt/Wasm: Add a number of Wasm-specific intrinsics and transients.
The motivation is mainly to get intrinsics for the bit-conversions
between integers and floating point numbers, as well as for
`numberLeadingZeros`. These are building blocks for many other
low-level operations, and their JS-builtin-based implementation is
really bad on Wasm for those use cases.

Once we have the infrastructure for those as transients in the Wasm
backend, we also take the opportunity to add a series of other
methods that have a direct Wasm opcode equivalent.