Wasm: Dedicated implementation of `Double.hashCode`.
The implementation based on `FloatingPointBits` is poor for a Wasm
output, as it uses JS interop. Given that we have an intrinsic for
`doubleToLongBits`, we can implement it much more efficiently for
Wasm.

This is important because this function is called to compute the
hash code of all JS `number`s, including those that originate as
`Int`s.