Use a typeclass to cut down on boilerplate in FunctionBuilder.
We still need an overload of each method for the empty block type,
but the other ones disappear.

Also, we cannot remove the boilerplate from the `switch` overloads,
since `switch` ultimately wants a pair of `FunctionType`s so that
it can decompose them. It could not do that with `BlockType`s.