Get rid of Phase 2 of the itable bucket assignment.
We observe that interfaces must all be spine types, and since we
only need to assign itable indices to interfaces, we can entirely
skip Phase 2 of the paper's algorithm.

We also switch to using direct *parents* (but including classes)
instead of the full interface ancestors to better match the paper's
algorithm.